Rename schedules to tasks, split up tasks view
